Finished Commit

+++ resources/views/data-management/book-pages/book.blade.php
@@ -30,25 +30,33 @@
                                 <div class="col-3 col-md-3 col-sm-auto  col-lg-3">
                                     <a href="{{ route('buku.show', $item->id_buku) }}" class="btn btn-primary">Detail</a>
                                 </div>
-                                <div class="col-5 col-md-4 col-sm-auto  col-lg-4">
-                                    <a href="/delete-buku/{{ $item->id_buku }}" class="btn btn-danger ms-2">Hapus</a>
+                                <div class="col-5 col-md-4 col-sm-auto col-lg-4">
+                                    @if (Auth::user()->role == 'admin' || Auth::user()->role == 'staff')
+                                        <a href="/delete-buku/{{ $item->id_buku }}" class="btn btn-danger ms-2">Hapus</a>
+                                    @endif
                                 </div>
-                                {{-- @if ($koleksi['id_buku'] == $item->id_buku) --}}
-                                <div class="col-6 col-md-5 col-sm-auto col-lg-5">
-                                    <form action="{{ route('koleksi.store') }}" method="post">
-                                        @csrf
-                                        <input type="text" name="id_user" value="{{ Auth::user()->id }}" hidden
-                                            id="">
-                                        <input type="text" name="id_buku" value="{{ $item->id_buku }}" hidden
-                                            id="">
-                                        <button type="submit" class="btn btn-info" data-bs-toggle="tooltip"
-                                            data-bs-placement="top" data-bs-custom-class="custom-tooltip"
-                                            data-bs-title="Add to personal collections">
-                                            <i class='bx bxs-star'></i>
-                                        </button>
-                                    </form>
-                                </div>
-                                {{-- @endif --}}
+                                <?php
+                                $koleksi = App\Models\KoleksiBuku::where('id_buku', $item->id_buku)
+                                    ->where('id_user', Auth::user()->id)
+                                    ->first();
+                                ?>
+                                @if (!$koleksi)
+                                    <div class="col-6 col-md-5 col-sm-auto col-lg-5">
+                                        <form action="{{ route('koleksi.store') }}" method="post">
+                                            @csrf
+                                            <input type="text" name="id_user" value="{{ Auth::user()->id }}" hidden
+                                                id="">
+                                            <input type="text" name="id_buku" value="{{ $item->id_buku }}" hidden
+                                                id="">
+                                            <button type="submit" class="btn btn-info" data-bs-toggle="tooltip"
+                                                data-bs-placement="top" data-bs-custom-class="custom-tooltip"
+                                                data-bs-title="Add to personal collections">
+                                                <i class='bx bxs-star'></i>
+                                            </button>
+                                        </form>
+                                    </div>
+                                @endif
+
                             </div>
                         </div>
                     </div>
